replaced lp_solver_type_t with existent method_t

+++ cpp/src/linear_programming/utilities/cython_solve.cu
@@ -182,7 +182,7 @@ linear_programming_ret_t call_solve_lp(
     solution.get_additional_termination_information().gap,
     solution.get_additional_termination_information().number_of_steps_taken,
     solution.get_additional_termination_information().solve_time,
-    (int)solution.get_additional_termination_information().solved_by};
+    solution.get_additional_termination_information().solved_by};
 
   return lp_ret;
 }
